fixed so step is a double not a float Original: ae035a513a199f8a42e5c7b00019357f2c6aef0f

+++ lalapps/src/calibration/mkcalfac.c
@@ -97,7 +97,7 @@ int main( int argc, char *argv[] )
 {
   static LIGOTimeGPS  tbeg;
   static size_t size;
-  static float  step;
+  static double step;
   static char   site;
   struct FrFile *frfile = NULL;
   struct FrameH *frame  = NULL;
